test: update tests for Grails 8 and SB4

+++ plugin-core/examples/functional-test-app/grails-app/views/testRole/show.gsp
@@ -41,10 +41,10 @@
 				</li>
 				</g:if>
 			</ol>
-			<g:form url="[resource:testRole, action:'delete']" method="DELETE">
+			<g:form url="[resource: testRole, action: 'delete']" method="DELETE" onsubmit="return confirm('Are you sure?');">
 				<fieldset class="buttons">
 					<g:link class="edit" action="edit" resource="${testRole}">Edit</g:link>
-					<g:actionSubmit class="delete" action="delete" value='Delete' onclick="return confirm('Are you sure?');" />
+					<input type="submit" class="delete" value="Delete">
 				</fieldset>
 			</g:form>
 		</div>
